requirements.txt
black==21.11b1
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
python-dotenv==0.19.2
regex==2021.11.10
tomli==1.2.2
typing-extensions==4.0.0
Werkzeug==2.0.2
